+++ README.md
@@ -3,7 +3,7 @@
 [<img alt="crates.io" src="https://img.shields.io/crates/v/discretionary_engine.svg?color=fc8d62&logo=rust" height="20" style=flat-square>](https://crates.io/crates/discretionary_engine)
 [<img alt="docs.rs" src="https://img.shields.io/badge/docs.rs-66c2a5?style=for-the-badge&labelColor=555555&logo=docs.rs&style=flat-square" height="20">](https://docs.rs/discretionary_engine)
 [<img alt="build status" src="https://img.shields.io/github/actions/workflow/status/valeratrades/discretionary_engine/ci.yml?branch=master&style=for-the-badge&style=flat-square" height="20">](https://github.com/valeratrades/discretionary_engine/actions?query=branch%3Amaster)
-![Lines Of Code](https://img.shields.io/badge/LoC-2609-lightblue)
+![Lines Of Code](https://img.shields.io/badge/LoC-2625-lightblue)
 
 Places and follows a position from a definition of _what the target position is_
